Move local storage helpers to hook file
