+++ src/commons/hooks/useLocalStorageState.ts
@@ -1,6 +1,13 @@
 import { useEffect, useState } from 'react';
 
-import { readLocalStorage, setLocalStorage } from '../utils/LocalStorageHelper';
+export const readLocalStorage = (key: string, defaultValue?: any) => {
+  const localStorageValue = window.localStorage.getItem(key);
+  return localStorageValue ? JSON.parse(localStorageValue) : defaultValue;
+};
+
+export const setLocalStorage = (key: string, value: any) => {
+  window.localStorage.setItem(key, JSON.stringify(value));
+};
 
 /**
  * This hook usage is similar to useState, the only difference
